% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_abbott.R
\name{corr_abbott}
\alias{corr_abbott}
\title{Correcion de Abbott para datos de mortalidad natural}
\usage{
corr_abbott(df, control = 0)
}
\arguments{
\item{df}{data frame}

\item{control}{Variable que indica como se codifica la columna control (Por ej.: 0, "control", etc)}
}
\value{
retorna el data frame corregido
}
\description{
Correcion de Abbott para datos de mortalidad natural
}
