% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{database_tox}
\alias{database_tox}
\title{Datos de ejemplo del paquete tox}
\format{
\subsection{\code{database_tox}}{

El dta frame contiene 636 filas y 9 columnas:
\describe{
\item{fecha}{Fecha de realización del ensayo.}
\item{especie}{Nombre científico de la especie estudiada.}
\item{n}{Número de individuos testeados.}
\item{exposicion_horas}{Duración del test en horas.}
\item{replica}{Número de replicas del ensayo.}
\item{dosis}{Concentracion de las dósis en ppb.}
\item{supervivencia}{Número de Individuos que sobrevivieron al test.}
\item{respuesta}{Número de Individuos que no sobrevivieron al test.}
}
}
}
\source{
\url{https://www.ilpla.edu.ar/}
}
\usage{
database_tox
}
\description{
Conjunto de datos de experimentos de toxicidad correspondientes a...
}
\keyword{datasets}
