% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probit.R
\name{probit}
\alias{probit}
\title{Funci칩n para an치lisis Probit con correcciones}
\usage{
probit(data, control_dosis = 0, conf.level = 0.95)
}
\arguments{
\item{data}{Dataframe con columnas "dosis" - "n" - "respuesta".}

\item{control_dosis}{Factor que representa el control si lo hay. Ej: 0, C, control}

\item{conf.level}{Nivel de confianza}
}
\value{
Una lista.
}
\description{
Funci칩n para an치lisis Probit con correcciones
}
\examples{
datos <- importar(data)
probit(datos)
}
